Split FirebaseManager to two files according to different methods functionality.
