+++ SoundScape/Model/Common/Common.swift
@@ -50,14 +50,7 @@ struct CommonUsage {
     static let font = "PingFang TC"
     static let fontSemibold = "PingFangTC-Semibold"
     
-    static let fakeMap = "fakeMap"
-    static let audioImage = "audioImage"
-    static let audioImage2 = "audioImage2"
-    static let profileCover = "profileCover"
-    static let profileCover2 = "profileCover2"
-    static let profileCover3 = "profileCover3"
     static let profileCover4 = "profileCover4"
-    static let profileCover5 = "profileCover5"
     static let yeh1024 = "yeh1024"
     static let launchScreen1 = "launchScreen1"
     
@@ -154,13 +147,22 @@ struct CommonUsage {
         static let loading = "loading..."
     }
     
-    struct CollectionName {
-        static let allUsers = "AllUsers"
-        static let allAudioFiles = "AllAudioFiles"
-        static let allLocations = "AllLocations"
-        static let comments = "Comments"
-    }
+}
+
+struct Constant {
+    static let allUsers = "AllUsers"
+    static let allAudioFiles = "AllAudioFiles"
+    static let allLocations = "AllLocations"
+    static let comments = "Comments"
+    static let followedBy = "followedBy"
+    static let following = "following"
+    static let myFavorite = "myFavorite"
+    static let profilePicture = "profilePicture"
+    static let blackList = "blackList"
+    static let userPic = "userPic"
+    static let coverPic = "coverPic"
     
+    static let createdTime = "createdTime"
 }
 
 extension UITextField {
